changes requested by Kai
